+++ build.sh.txt
@@ -1,3 +1,3 @@
 #!/bin/bash -xe
 mkdir -p target
-javac -d target/ src/*java src/operatingsystemsassignment/*java 
+javac -d target/ src/*java src/osone/*java
