+++ run.sh.txt
@@ -1,5 +1,5 @@
 #!/bin/bash -e
 
 cd target/
-java operatingsystemsassignment/Controller 
+java osone/Controller 
 cd ..
