+++ src/module-info.java
@@ -1,10 +1,10 @@
 /**
- * 
- */
-/**
- * @author azart
  *
  */
-module operatingsystemsassignment {
+/**
+ * @author myID_xxxxxxx
+ *
+ */
+module osone {
 	requires java.desktop;
 }
