--- src/operatingsystemsassignment/Settings.java
@@ -1,7 +1,0 @@
-package operatingsystemsassignment;
-public class Settings {
-    public class Parameters {
-        public static int queueCapacity = 5;
-        public static int sensorsCount = 3;
-    }
-}
