+++ src/osone/ConsoleStyles.java
@@ -1,6 +1,11 @@
-package operatingsystemsassignment;
+package osone;
 public class ConsoleStyles {
     // ANSI escape codes for used colors
+    public static final String resetBG = "\u001B[0m";
+    public static final String redBG = "\u001B[41m";
+    public static final String blackBG = "\u001B[40m";
+    public static final String yellowBG = "\u001B[43m";
+
     public static String infoColor = "\u001B[33;1m"; // darkYellowColor
     public static String errorColor = "\u001B[31m"; // redColor
     public static String greenColor = "\u001B[32m";
