+++ src/osone/Settings.java
@@ -1,0 +1,7 @@
+package osone;
+public class Settings {
+    public class Parameters {
+        public static int queueCapacity = 10;
+        public static int sensorsCount = 1;
+    }
+}
